Fix issue of overriding all settings when DeletelLastWelcomeMessage not set on exiting bots.
When the bot is started check if the 9 original group settings are set, if they are not all set then the default settings will be generated for the 9 original group settings as well as the new setting DeletelLastWelcomeMessage. For running bots and bots restarted with the original 9 group settings already set, the new DeleteLastWelcomeMessage group setting will only be added when the /welcome command is used.